requirements.txt
asyncio
aiohttp==3.9.1
websockets==12.0
psutil==5.9.6
netifaces==0.11.0
cryptography==41.0.7
pyyaml==6.0.1
click==8.1.7
colorama==0.4.6
rich==13.7.0
uvloop==0.19.0; sys_platform != "win32"
flask==3.0.0
flask-socketio==5.3.6
requests==2.31.0
dnspython==2.4.2
